data_catalog_event_handler/requirements.txt
fdk
oci
requests
pyyaml
